Move the logo generation codes to pycamp.py
