+++ pyconjpbot/plugins/pycamp.py
@@ -1,11 +1,15 @@
+import os
+from io import StringIO
 from datetime import timedelta, datetime
-
 from dateutil import parser
+
 from jira import JIRA, JIRAError
 from slackbot import settings
 from slackbot.bot import respond_to
+from slackbot.utils import create_tmp_file
 import requests
 from bs4 import BeautifulSoup
+from PIL import Image, ImageFont, ImageDraw
 
 from ..google_plugins.google_api import get_service
 from ..botmessage import botsend, botwebapi
@@ -44,6 +48,16 @@ SHEET_ID = '1LEtpNewhAFSf_vtkhTsWi6JGs2p-7XHZE8yOshagz0I'
 # 参加者枠の短い名前
 SHORT_PTYPE_NAMES = ('学生', 'TA', 'スタッフ',
                      '本イベント参加者', '懇親会のみ')
+
+# Settings for the logo generation.
+MAIN_COLOR = (90, 200, 233)
+TEXT_SIZE = 120
+TEXT_HEIGHT = 200
+FONT = 'RictyDiminished-Regular.ttf'
+IMAGES = (
+        ('pycamp_logo.png', (1080, 1080)),
+        ('pycamp_logo_horizontal.png', (2827, 1080)),
+)
 
 HELP = """
 `$pycamp create (地域) (開催日) (現地スタッフJIRA) (講師のJIRA)`: pycamp のイベント用issueを作成する
@@ -313,9 +327,37 @@ def pycamp_summary(message, party=None):
     botwebapi(message, attachements)
 
 
+@respond_to('^pycamp\s+logo\s+(\S+)')
+def pycamp_logo(message, title):
+    botsend(message, 'Python Boot Camp ロゴ作成中... :hammer:')
+
+    for name, size in IMAGES:
+        logo_image = Image.open(os.path.join('templates', name))
+        logo_image = logo_image.convert('RGBA')
+        logo_image.thumbnail(size)
+
+        width, height = size
+
+        background = Image.new('RGBA', (width, TEXT_HEIGHT), MAIN_COLOR)
+        font = ImageFont.truetype(os.path.join('fonts', FONT), size=TEXT_SIZE)
+        draw = ImageDraw.Draw(background)
+        text_width, _ = draw.textsize(title, font=font)
+        draw.text(((width - text_width) / 2, 0), title, font=font, fill=(0, 0, 0))
+
+        logo_image.paste(background, (0, height - TEXT_HEIGHT))
+
+        with create_tmp_file() as tmpf:
+            logo_image.save(tmpf, 'png')
+            message.channel.upload_file(name, tmpf)
+
+    botsend(message, '作成完了 :thumbsup:')
+
+
 @respond_to('^pycamp\s+help')
 def pycamp_help(message):
     """
     ヘルプメッセージを返す
     """
     botsend(message, HELP)
+
+
